requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
apsw==3.49.2.0
    # via apswutils
apswutils==0.0.2
    # via fastlite
asgiref==3.8.1
    # via daphne
attrs==25.3.0
    # via
    #   service-identity
    #   twisted
autobahn==24.4.2
    # via daphne
automat==25.4.16
    # via twisted
beautifulsoup4==4.13.4
    # via python-fasthtml
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via cryptography
click==8.2.1
    # via uvicorn
constantly==23.10.4
    # via twisted
cryptography==45.0.2
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==4.2.0
    # via -r requirements.in
fastcore==1.8.2
    # via
    #   apswutils
    #   fastlite
    #   python-fasthtml
fastlite==0.1.3
    # via python-fasthtml
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via python-fasthtml
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.10
    # via
    #   anyio
    #   httpx
    #   hyperlink
    #   twisted
incremental==24.7.2
    # via twisted
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via python-fasthtml
oauthlib==3.2.2
    # via python-fasthtml
packaging==25.0
    # via
    #   fastcore
    #   pytest
pluggy==1.6.0
    # via pytest
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pycairo==1.28.0
    # via
    #   pygobject
    #   toga-gtk
pycparser==2.22
    # via cffi
pygobject==3.52.3
    # via toga-gtk
pyopenssl==25.1.0
    # via twisted
pytest==8.3.5
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.26.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via python-fasthtml
python-dotenv==1.1.0
    # via uvicorn
python-fasthtml==0.12.18
    # via -r requirements.in
python-multipart==0.0.20
    # via python-fasthtml
pyyaml==6.0.2
    # via uvicorn
service-identity==24.2.0
    # via twisted
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
soupsieve==2.7
    # via beautifulsoup4
starlette==0.46.2
    # via python-fasthtml
toga==0.5.1
    # via -r requirements.in
toga-core==0.5.1
    # via toga-gtk
toga-gtk==0.5.1
    # via toga
travertino==0.5.1
    # via toga-core
twisted[tls]==24.11.0
    # via daphne
txaio==23.1.1
    # via autobahn
typing-extensions==4.13.2
    # via
    #   anyio
    #   beautifulsoup4
    #   pyopenssl
    #   twisted
uvicorn[standard]==0.34.2
    # via python-fasthtml
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn
zope-interface==7.2
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
